pyproject.toml
[tool.poetry]
name = "notebook"
version = "0.1.0"
description = "test project"
authors = ["Alexander Zhurilo <sashazhurilo@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.4"
djangorestframework = "^3.14.0"
drf-spectacular = "^0.26.4"
drf-yasg = "^1.21.7"
django-cors-headers = "^4.2.0"
drf-standardized-errors = "^0.12.5"
psycopg2 = "^2.9.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
